test: remove stale session header e2e
